Refactor authentication API; rename UserApi to AuthApi and update related references for improved clarity and consistency

+++ src/main/resources/db/migrations/V2025_02_05_14_00__create_db.sql
@@ -24,7 +24,7 @@ create TABLE AUTH.USERS (
     password VARCHAR(70),
     status TEXT,
     role_code TEXT,
-    version INTEGER,
+    version BIGINT,
     created_by TEXT,
     created_at TIMESTAMP WITHOUT TIME ZONE,
     modified_by TEXT,
